stm32f4xx: Move DmaClock into shared dma module
Instead of providing a Dma1Clock and a Dma2Clock, we can use a shared
class for both DMA's, as they share the exact same code.
